Moving activity type too
