+++ app/org/sagebionetworks/bridge/models/schedules/Activity.java
@@ -1,13 +1,12 @@
 package org.sagebionetworks.bridge.models.schedules;
 
-import static org.sagebionetworks.bridge.models.tasks.ActivityType.TASK;
-import static org.sagebionetworks.bridge.models.tasks.ActivityType.SURVEY;
+import static org.sagebionetworks.bridge.models.schedules.ActivityType.SURVEY;
+import static org.sagebionetworks.bridge.models.schedules.ActivityType.TASK;
 
 import java.util.Objects;
 
 import org.joda.time.DateTime;
 import org.sagebionetworks.bridge.models.BridgeEntity;
-import org.sagebionetworks.bridge.models.tasks.ActivityType;
 import org.sagebionetworks.bridge.validators.ActivityValidator;
 import org.sagebionetworks.bridge.validators.Validate;
