+++ app/org/sagebionetworks/bridge/models/schedules/ActivityType.java
@@ -1,0 +1,6 @@
+package org.sagebionetworks.bridge.models.schedules;
+
+public enum ActivityType {
+    SURVEY,
+    TASK
+}
