--- app/org/sagebionetworks/bridge/models/tasks/ActivityType.java
@@ -1,6 +1,0 @@
-package org.sagebionetworks.bridge.models.tasks;
-
-public enum ActivityType {
-    SURVEY,
-    TASK
-}
